Move tests of old code to correct package
COSDK-1119
